LayoutTests/imported/w3c/web-platform-tests/webxr/xrSession_input_events_end.https-expected.txt


FAIL Calling end during an input callback stops processing at the right time - webgl assert_equals: Expected end event, but got selectstart event instead expected "end" but got "selectstart"
FAIL Calling end during an input callback stops processing at the right time - webgl2 assert_equals: Expected end event, but got selectstart event instead expected "end" but got "selectstart"
